Conflicts:
	setup_dependencies.py
	src/couchapps/ReqMgr/updates/updaterequest.js
	src/python/WMCore/Lexicon.py
	src/python/WMCore/ReqMgr/DataStructs/RequestStatus.py
	src/python/WMCore/ReqMgr/Service/Request.py
	src/python/WMCore/WMSpec/StdSpecs/StdBase.py
	src/python/WMCore/WMSpec/WMWorkloadTools.py
	test/data/ReqMgr/requests/ReReco.json

+++ src/python/WMCore/Services/Requests.py
@@ -92,7 +92,7 @@ class Requests(dict):
                         idict.get('service_name')))
             self["cachepath"] = cache_dir
             self["req_cache_path"] = os.path.join(cache_dir, '.cache')
-        self.setdefault("timeout", 30)
+        self.setdefault("timeout", 300)
         self.setdefault("logger", logging)
 
         check_server_url(self['host'])
@@ -472,7 +472,6 @@ class Requests(dict):
             c.setopt(pycurl.CUSTOMREQUEST, 'PUT')
         else:
             raise HTTPException("Verb %s not sopported for upload." % verb)
-        print url
         c.setopt(c.URL, url)
         fullParams = [(fieldName, (c.FORM_FILE, fileName))]
         fullParams.extend(params)
@@ -582,4 +581,5 @@ class TempDirectory():
         self.dir = idir
 
     def __del__(self):
-        shutil.rmtree(self.dir, ignore_errors = True)
+        if shutil:
+            shutil.rmtree(self.dir, ignore_errors = True)
